Eliminate 'bass drop out' by disabling XRUN, and substituting silence
The 'bass drop out' effect is known to be triggered after recovering
from an XRUN condition[1].

By eliminating XRUNs, we prevent 'bass drop out'.

The quid quo pro is that, where an xrun would otherwise have occurred,
we will experience a small but noticeable glitch in the audio.

However it is not certain that occurrence of these xruns would actually
trigger drop out. I find, perhaps, 1 in 5, or 1 in 10.

Eliminating the possibility of xruns appears to cure the matter, but
will produce a small but noticeable glitch in the audio. So it becomes
a matter of user preference: Are a few unnecessary glitches a price
worth paying for eliminating 'bass drop out' ?

[1] For reasons unknown, we find that, after having apparently
successfully recovered from xrun (by executing 'snd_pcm_recover'), the
system occasionally switches the Left/Tweeter and Right/Woofer channels
around. This seems to happen well down the playback chain, perhaps at
the kernel driver/DMA level.

Sending Tweeter sound to the Woofer, and vice versa, accounts for the
observed 'bass drop out'. If listening through headphones, we hear that
the left and right channels are swapped.